Complete first go at refactoring to language agnostic tool
