+++ src/runcrate/constants.py
@@ -24,3 +24,8 @@ WORKFLOW_PROFILE = f"{WORKFLOW_PROFILE_BASE}/{PROFILES_VERSION}"
 PROVENANCE_PROFILE = f"{PROVENANCE_PROFILE_BASE}/{PROFILES_VERSION}"
 
 TERMS_NAMESPACE = "https://w3id.org/ro/terms/workflow-run"
+
+WROC_PROFILE_VERSION = "1.0"
+DOCKER_IMG_TYPE = "https://w3id.org/ro/terms/workflow-run#DockerImage"
+
+
