+++ tests/test_step_mapping.py
@@ -26,8 +26,10 @@ def converter():
 def test_step_maps_cwl(data_dir, converter):
     wf_basename = "exome-alignment-packed.cwl"
     wf_path = data_dir / wf_basename
-    cwl_defs = converter.get_workflow(wf_path)
-    step_maps = converter.get_step_maps(cwl_defs)
+    converter.wf_path = wf_path
+    cwl_defs = converter.get_workflow()
+    converter.workflow_definition = cwl_defs
+    step_maps = converter.get_step_maps()
     assert set(step_maps) == {wf_basename}
     sm = step_maps[wf_basename]
     assert len(sm) == 8
@@ -48,9 +50,10 @@ def test_step_maps_cwl(data_dir, converter):
 
 
 def test_step_maps_disconnected_cwl(data_dir, converter):
-    wf_path = data_dir / "no-output-run-1/workflow/packed.cwl"
-    cwl_defs = converter.get_workflow(wf_path)
-    step_maps = converter.get_step_maps(cwl_defs)
+    converter.wf_path = data_dir / "no-output-run-1/workflow/packed.cwl"
+    cwl_defs = converter.get_workflow()
+    converter.workflow_definition = cwl_defs
+    step_maps = converter.get_step_maps()
     assert set(step_maps) == {"packed.cwl"}
     sm = step_maps["packed.cwl"]
     assert set(sm) == {"main/date_step", "main/echo_step", "main/date2_step"}
